// server/sonar-web/src/main/js/types/appstate.ts
/*
 * SonarQube
 * Copyright (C) 2009-2025 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */

import { EditionKey } from './editions';
import { GlobalSettingKeys } from './settings';
import { Extension } from './types';

export interface AppState {
  authenticationError?: boolean;
  authorizationError?: boolean;
  canAdmin?: boolean;
  documentationUrl: string;
  edition?: EditionKey;
  globalPages?: Extension[];
  instanceUsesDefaultAdminCredentials?: boolean;
  needIssueSync?: boolean;
  productionDatabase: boolean;
  qualifiers: string[];
  settings: { [key in GlobalSettingKeys]?: string };
  standalone?: boolean;
  version: string;
  versionEOL: string;
  webAnalyticsJsPath?: string;
}